+++ ui/src/ui_drawable.h
@@ -11,7 +11,7 @@ class UIDrawable {
  public:
   virtual void Render(RenderWindow *window) = 0;
 
-  [[nodiscard]] virtual bool CheckClicked(Vector2i coords) = 0;
+  [[nodiscard]] virtual bool CheckClicked(Vector2f coords) = 0;
 
   [[nodiscard]] virtual bool IsClickable() = 0;
 };
